+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [6.9.1] - Unreleased
+## [6.9.1] - 2022-05-17
 
 ### Fixed
 - Text message on `Payment Methods > RDFa` tab
@@ -1077,7 +1077,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [6.0-beta.1] - 2016-11-30
 
-[6.9.1]: https://github.com/OXID-eSales/oxideshop_ce/compare/v6.9.0...b-6.3.x
+[6.9.1]: https://github.com/OXID-eSales/oxideshop_ce/compare/v6.9.0...v6.9.1
 [6.9.0]: https://github.com/OXID-eSales/oxideshop_ce/compare/v6.8.0...v6.9.0
 [6.8.0]: https://github.com/OXID-eSales/oxideshop_ce/compare/v6.7.1...v6.8.0
 [6.7.2]: https://github.com/OXID-eSales/oxideshop_ce/compare/v6.7.1...v6.7.2